Added charts handling on the backend and changed the structure a bit
