+++ ainalytics/src/ainalytics/app.py
@@ -2,8 +2,8 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
 
-from .schemas import ChartDisplay, ChartDisplayId, Message
-from .service import run_new_agent, run_agent
+from ainalytics.agent.router import router as agent_router
+from ainalytics.charts.router import router as charts_router
 
 app = FastAPI()
 
@@ -22,14 +22,5 @@ def configure_cors():
 configure_cors()
 
 
-@app.post("/agent", response_model=ChartDisplayId)
-@app.post("/agent/new", response_model=ChartDisplayId)
-def agent_new(request: Message):
-    state, data, obj_id = run_new_agent(request.message)
-    return ChartDisplayId(type=state.chart, data=data, query=state.query, id=obj_id)
-
-
-@app.post("/agent/{id}", response_model=ChartDisplay)
-def agent(request: Message, id: int):
-    state, data = run_agent(request.message, id)
-    return ChartDisplay(type=state.chart, data=data, query=state.query)
+app.include_router(agent_router)
+app.include_router(charts_router)
